+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Added: Browser utilities `CConnoisseur.ascend` and `CConnoisseur.reincarnate`.
+- Added: Browser utility `CConnoisseur.redrawBankMinigame`.
 
 ## 0.2.0 - 2021-10-01
 - Added: Explicit support for Firefox.